Code cleanup after review.
