+++ src/main/kotlin/io/perfometer/runner/ThreadPoolScenarioRunner.kt
@@ -1,30 +1,25 @@
 package io.perfometer.runner
 
+import io.perfometer.dsl.HttpStep
 import io.perfometer.dsl.PauseStep
 import io.perfometer.dsl.RequestStep
-import io.perfometer.dsl.HttpStep
 import io.perfometer.http.client.HttpClient
-import io.perfometer.statistics.ConcurrentQueueScenarioStatistics
 import io.perfometer.statistics.PauseStatistics
-import io.perfometer.statistics.RequestStatistics
-import io.perfometer.statistics.ScenarioStatistics
 import java.time.Duration
-import java.time.Instant
 import java.util.concurrent.CompletableFuture
 import java.util.concurrent.ExecutorService
 import java.util.concurrent.Executors
 import java.util.concurrent.TimeUnit
 
 internal class ThreadPoolScenarioRunner(
-        private val httpClient: HttpClient,
-        private val scenarioStatistics: ScenarioStatistics = ConcurrentQueueScenarioStatistics(Instant.now())
-) : ScenarioRunner {
+        httpClient: HttpClient,
+) : BaseScenarioRunner(httpClient) {
 
-    override fun runUsers(userCount: Int, block: () -> Unit) {
+    override fun runUsers(userCount: Int, action: () -> Unit) {
         val scenarioExecutor = Executors.newFixedThreadPool(userCount)
         CompletableFuture.allOf(
                 *(0 until userCount)
-                        .map { CompletableFuture.runAsync(block, scenarioExecutor) }
+                        .map { CompletableFuture.runAsync(action, scenarioExecutor) }
                         .toTypedArray())
                 .join()
         shutdown(scenarioExecutor)
@@ -37,10 +32,6 @@ internal class ThreadPoolScenarioRunner(
         }
     }
 
-    override fun statistics(): ScenarioStatistics {
-        return scenarioStatistics
-    }
-
     private fun shutdown(scenarioExecutor: ExecutorService) {
         scenarioExecutor.shutdown()
         val isTerminated = scenarioExecutor.awaitTermination(10, TimeUnit.SECONDS)
@@ -51,20 +42,6 @@ internal class ThreadPoolScenarioRunner(
 
     private fun pauseFor(duration: Duration) {
         Thread.sleep(duration.toMillis())
-        scenarioStatistics.gather(PauseStatistics(duration))
-    }
-
-    private fun executeHttp(requestStep: RequestStep) {
-        val startTime = Instant.now()
-        val request = requestStep.request
-        val response = httpClient.executeHttp(request)
-        val timeElapsed = Duration.between(startTime, Instant.now())
-        request.consumer(response)
-        scenarioStatistics.gather(RequestStatistics(
-                request.method,
-                request.pathWithParams,
-                timeElapsed,
-                response.status)
-        )
+        statistics.gather(PauseStatistics(duration))
     }
 }
